feat(like-mobile): GridList 추가 및 컨테이너 정리 (refactor/web 동기화)
